+++ packages/message-port-rpc/src/index.ts
@@ -1,3 +1,4 @@
+import forGenerator from './forGenerator';
 import messagePortRPC from './messagePortRPC';
 
-export { messagePortRPC };
+export { forGenerator, messagePortRPC };